Ajout des principales fonctionnalitÃs
- FonctionnalitÃs :
  -> Attaquer un autre serviteur :
     Pour l'attaque, on crÃe une nouvelle action "attaque" avec pour dÃgÃts
     les dÃgÃts du serviteur.
  -> Lancer l'action spÃciale du hÃros
  -> Piocher une nouvelle carte
  -> Jouer une carte (serviteur ou sort)
  -> Communication avec les interfaces Repositories (Spring jpa)

- Ajout de tests unitaires pour les fonctionnalitÃs listÃes ci-dessus

- A faire :
  -> Interface pour communiquer avec le client (websocket)
  -> Matchmaker
  -> Tests unitaires pour les classes Joueur et Jeu

+++ src/main/java/fr/univ_nantes/slightstone/model/ActionsRepository.java
@@ -2,6 +2,4 @@ package fr.univ_nantes.slightstone.model;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionsRepository extends CrudRepository<Action, Integer> {
-
-}
+public interface ActionsRepository extends CrudRepository<Action, Long> {}
